refactored and added utils for summary
